+++ CHANGELOG.md
@@ -2,3 +2,8 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 This change log adheres to standards from [Keep a CHANGELOG](http://keepachangelog.com).
+
+
+## Unreleased
+### Added
+* Add group-object-shorthands rule